migrate code monitors to enterprise
